iq2000: fix trap.h errors with gcc-15

+++ libgloss/iq2000/gettime.c
@@ -8,5 +8,5 @@
 int
 _gettimeofday (struct timeval *tp, void *tzp)
 {
-  return TRAP0 (SYS_gettimeofday, tp, tzp, 0);
+  return TRAP0 (SYS_gettimeofday, (unsigned long) tp, (unsigned long) tzp, 0);
 }
